Re-implement TheSecurity's refactors
