+++ src/Youtube2Mp3.Core/Extensions/TrackFormatExtensions.cs
@@ -17,10 +17,9 @@ namespace Youtube2Mp3.Core.Extensions
         {
             var author = track.Authors.FirstOrDefault();
 
-            if (string.IsNullOrEmpty(author))
-                return $"{track.Title}".RemoveIllegalPathCharacters();
-            else
-                return $"{author} - {track.Title}".RemoveIllegalPathCharacters();
+            return string.IsNullOrEmpty(author)
+                ? $"{track.Title}".RemoveIllegalPathCharacters()
+                : $"{author} - {track.Title}".RemoveIllegalPathCharacters();
         }
     }
 }
