+++ src/Youtube2Mp3.Spotify/Extensions/SpotifyClientExtensions.cs
@@ -6,20 +6,19 @@ namespace Youtube2Mp3.Spotify.Extensions
 {
     public static class SpotifyClientExtensions
     {
-        private static Regex spotifyPlaylistIdPattern = new Regex(@"playlist[\/|:](.{22})", RegexOptions.Compiled);
+        private static readonly Regex spotifyPlaylistIdPattern = new Regex(@"playlist[\/|:](.{22})", RegexOptions.Compiled);
 
         public static FullPlaylist? GetPlaylistByUrl(this SpotifyWebAPI api, string url)
         {
             var id = ParseSpotifyIdFromUrl(url);
-            if (id is null) { return null; }
-
-            return api.GetPlaylist(id);
+            return id is null
+                ? null
+                : api.GetPlaylist(id);
         }
 
         public static string? ParseSpotifyIdFromUrl(string url)
-        {
-            if (url is null || !spotifyPlaylistIdPattern.IsMatch(url)) { return null; }
-            return spotifyPlaylistIdPattern.Match(url).Groups[1].Value;
-        }
+            => url is null || !spotifyPlaylistIdPattern.IsMatch(url) 
+                ? null 
+                : spotifyPlaylistIdPattern.Match(url).Groups[1].Value;
     }
 }
